Split FK solver into files
